% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-bar-press.R
\name{bar_press}
\alias{bar_press}
\title{Barometric/total pressure (p)}
\usage{
bar_press(alt = 0)
}
\arguments{
\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of barometric pressure [kPa].
}
\description{
Calculates barometric pressure based on altitude. This is also called total
 pressure.
Eq. 3 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
bar_press()

bar_press(alt = 1000)
}
\seealso{
Other PsychrometricGraph: \code{\link{dewpoint}},
  \code{\link{enthalpy_intersect}}, \code{\link{enthalpy}},
  \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{my_round}}, \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
