% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaDBConenctor.R
\name{dbInit}
\alias{dbInit}
\title{EDAS DB Initialization}
\usage{
dbInit(username, passwd, dbName = "edas", server = "58.224.211.82",
  serverPort = 13306)
}
\arguments{
\item{username}{EDAS DataBase User Name}

\item{passwd}{EDAS DataBase Password}

\item{dbName}{EDAS DataBase Name}

\item{server}{DataBase Server IP Addrerss}

\item{serverPort}{DataBase Server Port}
}
\value{
EDAS DB Connection, Tables and Attributes
}
\description{
EDAS DB Initialization
}
\examples{
dbInit("EDASTest","EDASPW")
dbInit("EDASTest","EDASPW",  dbName="edas", server="58.224.211.82", serverPort=13306)
}
