% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-table.R
\name{describe.table}
\alias{describe.table}
\title{EDAS DB Integration API Interface - EDAS DB Table Description 
테이블 스키마 정보}
\usage{
\method{describe}{table}(edas, table)
}
\arguments{
\item{table}{테이블 명}

\item{connector}{EDAS.DB.API}
}
\value{
tableInfo - 테이블 스키마 
sql - 쿼리
}
\description{
EDAS DB Integration API Interface - EDAS DB Table Description 
테이블 스키마 정보
}
\examples{
db_server <- "10.0.0.1" 
db_port <- 13306
db_username <- "test"
db_password <- "xxxx"
db_name <- "edas"
dbConnector <- edas.db(testuser_name, testuser_password, "edas", testpoc_server, testpoc_port)
describe.table(dbConnector, "EDAS_S_USAGE")
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{edas.newModel}}, \code{\link{edas.validate}},
  \code{\link{get.table.sample}}
}
