% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-dewpoint.R
\name{dewpoint}
\alias{dewpoint}
\title{Dew-point temperature (t_d)}
\usage{
dewpoint(hum.ratio, alt = 0)
}
\arguments{
\item{hum.ratio}{Vector of humidity ratios [kg/kg].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of dew-point temperatures [degC].
}
\description{
Calculates the dew-point temperature based on humidity ratios.
Eq. 37 & 38 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
dewpoint(hum.ratio = 0.010)

hum.ratio = seq(0, 0.050, by = 0.001)
plot(dewpoint(hum.ratio = hum.ratio), hum.ratio,
     xlab = "Dew-point [degC]",
     ylab = "Humidity ratio [kg/kg]")
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{enthalpy_intersect}}, \code{\link{enthalpy}},
  \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{my_round}}, \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
