% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edas-datapreprocessing-daytype.R
\name{edas.datapreprocessing.daytype}
\alias{edas.datapreprocessing.daytype}
\title{EDAS Data Preprocess - Working/Non-Working/Special Days
데이터를 주중/주말/특별한 날로 구분}
\usage{
edas.datapreprocessing.daytype(dataset, datetimeCol, workingdays = c(1, 2, 3,
  4, 5), specialdays = NULL)
}
\arguments{
\item{dataset}{데이터 프레임 (data.frame) 형식의 input 데이터 셋}

\item{datetimeCol}{POSIXCT 타입의 날짜 시간 값을 나타내는 컬럼 명}

\item{workingdays}{주중을 정의 하는 요일들(1~7) numeric list, 기본값으로 월요일부터 금요일 (1~5)}

\item{specialdays}{Optional 특별한 날을 정의 하는 character list, 날짜 포맷 yyyy-mm-dd}
}
\value{
indextable - 데이터 프레임 (type, value), 주중/주말/특별한 날로 데이터가 어떤 값으로 구분되어있는지를 보여줌  
workingdays - list 형식의 업무요일 값(들) 
nonworkingdays - list 형식의 비업무요일 값(들)
specialdays - list 형식의 특별한 날짜(들) 
eval - input 데이터 셋(데이터 프레임)에 요일(dayOfweek) 값과 주중/주말/특별한 날을(datetype) 나타내는 값을 
추가한 결과
}
\description{
EDAS Data Preprocess - Working/Non-Working/Special Days
데이터를 주중/주말/특별한 날로 구분
}
\examples{
#generate a sample dataset
dataset <- data.frame("Datetime"=seq(ISOdate(1900,1,1), ISOdate(1901,1,1), length.out=20),
"x"=seq(1,10,length.out=20))

edas.datapreprocessing.daytype(dataset, "Datetime")

specialdays <- c("2016-01-02", "2016-01-03")
edas.datapreprocessing.daytype(dataset, "Datetime", specialdays = specialdays )
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{describe.table}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{edas.newModel}}, \code{\link{edas.validate}},
  \code{\link{get.table.sample}}
}
