% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edas-datavalidation-api-timeinterval.R
\name{edas.datavalidation.api.timeinterval}
\alias{edas.datavalidation.api.timeinterval}
\title{EDAS Data Validation Method - Time Intervals  
데이터 수집 주기 추출  
데이터의 시간 컬럼(datetimeColumn) 값을 기준으로 주기를 계산  
그룹 컬럼(groupByColumn) 값이 주여졌을 경우 groupByColumn 값 기준으로 
데이터를 나눈후 주기를 계산}
\usage{
edas.datavalidation.api.timeinterval(dataset, dateTimeColumn = "DateTime",
  groupByColumn = NA)
}
\arguments{
\item{dataset}{DataFrame}

\item{dateTimeColumn}{DateTime Column (requires POSIXCT Type)}

\item{groupByColumn}{Group by Column (Optional)}
}
\value{
units - time interval unit
stats - descriptive statistics of the estimated time intervals of the dataset or the groups 
dataset - the dataset with the evaluated time intervals, which is added as "Interval"
}
\description{
EDAS Data Validation Method - Time Intervals  
데이터 수집 주기 추출  
데이터의 시간 컬럼(datetimeColumn) 값을 기준으로 주기를 계산  
그룹 컬럼(groupByColumn) 값이 주여졌을 경우 groupByColumn 값 기준으로 
데이터를 나눈후 주기를 계산
}
\examples{
edas.datavalidation.api.timeinterval(df,"DateTime")
edas.datavalidation.api.timeinterval(df,"DateTime", "ID")

dataset <- data.frame("DateTime"=seq(ISOdate(1910,1,1), ISOdate(1999,1,1), length.out=10),
"ID"=letters[seq(from = 1, to =10)])
edas.datavalidation.api.timeinterval(dataset, "DateTime")


dataset2 <- data.frame("DateTime"=seq(ISOdate(1911,1,1), ISOdate(2000,1,1), length.out=10),
                      "ID"=letters[seq(from = 1, to =10)])
                      mergedDF <- rbind(dataset, dataset2)
                      edas.datavalidation.api.timeinterval(mergedDF, "DateTime", "ID")
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{edas.initmodel}},
  \code{\link{get.table.sample}}, \code{\link{modelInit}},
  \code{\link{validate.edas.data}}
}
\keyword{internal}
