% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAS-DB-API.R
\docType{class}
\name{EDAS.DB.API-class}
\alias{EDAS.DB.API-class}
\alias{EDAS.DB.API}
\title{EDAS DB API Class 
EDAS.DB.API Class
EDAS DB 접속 정보 관리}
\description{
EDAS DB API Class 
EDAS.DB.API Class
EDAS DB 접속 정보 관리
}
\seealso{
Other EDAS: \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{edas.initmodel}},
  \code{\link{get.table.sample}}, \code{\link{modelInit}},
  \code{\link{validate.edas.data}}
}
