% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edas-db-api-createConnection.R
\name{edas.db.api.createConnection}
\alias{edas.db.api.createConnection}
\title{EDAS DB Establish Connection 
EDAS DB 연결}
\usage{
edas.db.api.createConnection(obj)
}
\arguments{
\item{obj}{EDAS.DB.API}
}
\value{
Database Connection
}
\description{
EDAS DB Establish Connection 
EDAS DB 연결
}
\examples{
db_server <- "10.0.0.1"
db_port <- 13306
db_username <- "test"
db_password <- "xxxx"
db_name <- "edas"
dbConnector <- edas.db(db_username, db_password, db_name, db_server, db_port)
edas.db.api.createConnection(dbConnector)
}
\keyword{internal}
