% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edas-db-api-query.R
\name{edas.db.api.query}
\alias{edas.db.api.query}
\title{EDAS DB Integration API Interface - EDAS DB Query Executor
EDAS 데이터베이스 쿼리 실행 API}
\usage{
edas.db.api.query(connector, sql)
}
\arguments{
\item{connector}{EDAS.DB.API}

\item{sql}{a SQL query statement - 쿼리}
}
\value{
list - 쿼리 결과
}
\description{
EDAS DB Integration API Interface - EDAS DB Query Executor
EDAS 데이터베이스 쿼리 실행 API
}
\examples{
db_server <- "10.0.0.1"
db_port <- 13306
db_username <- "test"
db_password <- "xxxx"
db_name <- "edas"
dbConnector <- edas.db(db_username, db_password, db_name, db_server, db_port)
edas.db.api.query(dbConnector,"SELECT * FROM Test")

}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db}}, \code{\link{edas.initmodel}},
  \code{\link{get.table.sample}}, \code{\link{modelInit}},
  \code{\link{validate.edas.data}}
}
\keyword{internal}
