% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edas-db.R
\name{edas.db}
\alias{edas.db}
\title{EDAS DB Integration API Interface - EDAS DB Connection Interface
EDAS DB 접속 인터페이스}
\usage{
edas.db(username, password, database, host, port)
}
\arguments{
\item{username}{EDAS DataBase User Name - 접속정보 계정명}

\item{host}{DataBase Server IP Addrerss - EDAS DataBase 서버 정보}

\item{port}{DataBase Server Port - EDAS Database 서버 포트 정보}

\item{passwd}{EDAS DataBase Password - 접속정보 계정비번}

\item{dbName}{EDAS DataBase Name - 접속 DB명}
}
\value{
EDAS.DB.API
}
\description{
EDAS DB Integration API Interface - EDAS DB Connection Interface
EDAS DB 접속 인터페이스
}
\examples{
db_server <- "10.0.0.1" 
db_port <- 13306
db_username <- "test"
db_password <- "xxxx"
db_name <- "edas"
dbConnector <- edas.db(db_username, db_password, db_name, db_server, db_port)
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{describe.table}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}},
  \code{\link{edas.newModel}}, \code{\link{edas.validate}},
  \code{\link{get.table.sample}}
}
