% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDASModelAPI.R
\name{edas.initmodel}
\alias{edas.initmodel}
\title{EDAS Analysis Model - Show Message}
\usage{
edas.initmodel()
}
\description{
EDAS Analysis Model - Show Message
}
\examples{
initModels()
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{get.table.sample}}, \code{\link{modelInit}},
  \code{\link{validate.edas.data}}
}
