% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDASModelAPI.R
\name{edas.newModel}
\alias{edas.newModel}
\title{EDAS Analysis Model API - EDAS Analysis Model Initialization Method}
\usage{
edas.newModel()
}
\description{
EDAS Analysis Model API - EDAS Analysis Model Initialization Method
}
\examples{
edas.newModel()
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{describe.table}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{edas.validate}},
  \code{\link{get.table.sample}}
}
