% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edas-validate.R
\name{edas.validate}
\alias{edas.validate}
\title{EDAS Data Validation Methods}
\usage{
edas.validate(dataset, dateTimeColumn = NULL, groupByColumn = NULL,
  type = c("timeinterval", "daytype", "timetype"), ...)
}
\arguments{
\item{dataset}{a dataframe}

\item{dateTimeColumn}{(optional) a datetime format, giving datetimeCol is not in POSIXCT type, 
the function first tries to convert "datetimeCol" to POSIXCT type.}

\item{groupByColumn}{(optional) a group by column to create subsets which the function is applied individually.}

\item{type}{a list of a validation method to apply on the dataset}
}
\value{
units, stats, hist, dataset
}
\description{
EDAS Data Validation Methods
}
\examples{
edas.validate(dataset, "DATE")
edas.validate(dataset, "DATE", "ID")
edas.validate(dataset, "DATE", "ID", type="daytype")
edas.validate(dataset, "DATE", "ID", c("daytype","timetype"))
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{describe.table}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{edas.newModel}},
  \code{\link{get.table.sample}}
}
