% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-enthalpy.R
\name{enthalpy}
\alias{enthalpy}
\title{Enthalpy (h)}
\usage{
enthalpy(temp.db, hum.ratio)
}
\arguments{
\item{temp.db}{Vector of dry bulb temperatures [degC].}

\item{hum.ratio}{Vector of humidity ratios [kg/kg].}
}
\value{
Returns a vector of enthalpies [kJ/kg].
}
\description{
Calculates the enthalpy of a given moist air state.
Eq. 32 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
enthalpy(25, 0.010)

# Calculating cooling demand
density <- 1.2 # density of air (kg/m3)
(enthalpy(35, 0.020) - enthalpy(15, 0.010)) * density
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy_intersect}},
  \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{my_round}}, \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
