% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-enthalpy-intersect.R
\name{enthalpy_intersect}
\alias{enthalpy_intersect}
\title{Intersection between enthalpy line and 100 \% RH line}
\usage{
enthalpy_intersect(enthalpy, alt = 0)
}
\arguments{
\item{enthalpy}{Vector of enthalpies [kJ/kg].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns the dry-bulb temperature at which the enthalpy and
 100 \% RH line intersects [degC].
}
\description{
Calculates the intersection of RH 100 \% line and the enthalpy line.
}
\examples{
enthalpy_intersect(enthalpy = 50)
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy}},
  \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{my_round}}, \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
\keyword{internal}
