% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-table-sample.R
\name{get.table.sample}
\alias{get.table.sample}
\title{EDAS DB Integration API Interface - Data Sampling
DB 테이블 데이터 샘플링}
\usage{
get.table.sample(edasDBObj, table, samplingrate)
}
\arguments{
\item{edasDBObj}{EDAS.DB.API}

\item{table}{EDAS DataBase Table Name - 테이블 명}

\item{samplingrate}{a sampling ratio ( < 1 ) or number ( >= 1) - 샘플 데이터 건수 또는 비율}
}
\value{
sample data from the specified table
}
\description{
EDAS DB Integration API Interface - Data Sampling
DB 테이블 데이터 샘플링
}
\examples{
db_server <- "10.0.0.1" 
db_port <- 13306
db_username <- "test"
db_password <- "xxxx"
db_name <- "edas"
dbConnector <- edas.db(testuser_name, testuser_password, "edas", testpoc_server, testpoc_port)
get.table.sample(dbConnector, "EDAS_S_USAGE", .0002)
get.table.sample(dbConnector, "EDAS_S_USAGE", 1) 
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{describe.table}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{edas.newModel}}, \code{\link{edas.validate}}
}
