% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-hum-ratio-enthalpy.R
\name{hum_ratio_enthalpy}
\alias{hum_ratio_enthalpy}
\title{Humidity ratio for enthalpy lines}
\usage{
hum_ratio_enthalpy(enthalpy, temp.db, alt = 0)
}
\arguments{
\item{enthalpy}{Vector of enthalpies [kJ/kg].}

\item{temp.db}{Vector of dry bulb temperatures [degC].}
}
\value{
Returns a vector of humidity ratios [kg/kg].
}
\description{
Calculates the humidity ratio based on the enthalpy.
Eq. 32 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
hum_ratio_enthalpy(enthalpy = 30, temp.db = 25)
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy_intersect}},
  \code{\link{enthalpy}}, \code{\link{hum_ratio_rel_hum}},
  \code{\link{hum_ratio}}, \code{\link{my_round}},
  \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
\keyword{internal}
