% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-hum-ratio-rel-hum.R
\name{hum_ratio_rel_hum}
\alias{hum_ratio_rel_hum}
\title{Humidity ratio (W)}
\usage{
hum_ratio_rel_hum(rel.hum, temp.air, hum.ratio.max, alt = 0)
}
\arguments{
\item{rel.hum}{Vector of relative humidities [\%].}

\item{temp.air}{Vector of air temperatures [degC].}

\item{hum.ratio.max}{Max humidity ratio on psychrometric chart [kg/kg].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of humidity ratios [kg/kg].
}
\description{
Calculates the humidity ratio from relative humidity and temperature.
Eq. 25 & 12 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
hum_ratio_rel_hum(rel.hum = 60, temp.air = 25, hum.ratio.max = 0.02)
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy_intersect}},
  \code{\link{enthalpy}}, \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio}}, \code{\link{my_round}},
  \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
\keyword{internal}
