% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDASModelAPI.R
\name{modelInit}
\alias{modelInit}
\title{EDAS Analysis Model API - EDAS Analysis Model Initialization Method}
\usage{
modelInit()
}
\description{
EDAS Analysis Model API - EDAS Analysis Model Initialization Method
}
\examples{
modelInit()
}
\seealso{
Other EDAS: \code{\link{EDAS.DB.API-class}},
  \code{\link{edas.datapreprocessing.daytype}},
  \code{\link{edas.datavalidation.api.timeinterval}},
  \code{\link{edas.db.api.query}}, \code{\link{edas.db}},
  \code{\link{edas.initmodel}},
  \code{\link{get.table.sample}},
  \code{\link{validate.edas.data}}
}
