% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-my-round.R
\name{my_round}
\alias{my_round}
\title{Round to nearest number}
\usage{
my_round(x, n, type)
}
\arguments{
\item{x}{Vector of numbers to round.}

\item{n}{Nearest number to which the x-value should be rounded.}

\item{type}{A string (either 'ceiling' or 'floor') to indicate if the
number should be rounded up or down.}
}
\value{
Returns a vector of rounded numbers.
}
\description{
Round up or down to nearest user specified number.
}
\examples{
my_round(24.9, 5, "floor")
my_round(c(24.9, 20.1, 25), 5, "ceiling")
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy_intersect}},
  \code{\link{enthalpy}}, \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
\keyword{internal}
