% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-par-w-press.R
\name{par_w_press}
\alias{par_w_press}
\title{Partial water vapour pressure (p_w)}
\usage{
par_w_press(hum.ratio, alt = 0)
}
\arguments{
\item{hum.ratio}{Vector of humidity ratios [kg/kg].}

\item{alt}{Vector of altitudes [m]. Defaults to 0 m (sea level).}
}
\value{
Returns a vector of partial water vapour pressure [kPa].
}
\description{
Calculates partial water vapour pressure.
Eq. 36 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
par_w_press(hum.ratio = 0.010)

hum.ratio <- rep(c(0.005, 0.010), 2)
alt <- rep(c(0, 500), each = 2)
par_w_press(hum.ratio, alt)
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy_intersect}},
  \code{\link{enthalpy}}, \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{my_round}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
