% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-sat-w-press.R
\name{sat_w_press}
\alias{sat_w_press}
\title{Satuated water vapour pressure (p_ws)}
\usage{
sat_w_press(temp.air)
}
\arguments{
\item{temp.air}{Vector of air temperatures [degC].}
}
\value{
Returns a vector of satuated water pressure over water and
 ice [kPa].
}
\description{
Calculates the satuated water vapour pressure over water and ice.
Eq.5 & 6 - ASHRAE Fundamentals Handbook 2002, Psychrometrics.
}
\examples{
sat_w_press(temp.air = 25)
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy_intersect}},
  \code{\link{enthalpy}}, \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{my_round}}, \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}},
  \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
