% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-theme-climateeng-psy.R
\name{theme_climateeng_psy}
\alias{theme_climateeng_psy}
\title{Custom ggplot2}
\usage{
theme_climateeng_psy(asp = NULL)
}
\arguments{
\item{asp}{Aspect ratio of plot. Defaults to NULL.}
}
\value{
Returns a ggplot2 theme.
}
\description{
Custom theme for climateeng psychrometric chart.
}
\examples{
theme_climateeng_psy()
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy_intersect}},
  \code{\link{enthalpy}}, \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{my_round}}, \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}}, \code{\link{theme_parms}}
}
\author{
Christoffer Rasmussen
}
\keyword{internal}
