% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric-graph-theme-parms.R
\name{theme_parms}
\alias{theme_parms}
\title{ggplot2 parameters}
\usage{
theme_parms()
}
\value{
Returns a list of ggplot2 theme parameters.
}
\description{
Parameters for climateeng themes.
}
\examples{
theme_parms()
}
\seealso{
Other PsychrometricGraph: \code{\link{bar_press}},
  \code{\link{dewpoint}}, \code{\link{enthalpy_intersect}},
  \code{\link{enthalpy}}, \code{\link{hum_ratio_enthalpy}},
  \code{\link{hum_ratio_rel_hum}}, \code{\link{hum_ratio}},
  \code{\link{my_round}}, \code{\link{par_w_press}},
  \code{\link{psychrometric_chart}},
  \code{\link{sat_hum_ratio}}, \code{\link{sat_w_press}},
  \code{\link{slope_rel_hum}},
  \code{\link{theme_climateeng_psy}}
}
\author{
Christoffer Rasmussen
}
\keyword{internal}
